Moved out code from download to serve method
